@@ -1,40 +1,43 @@
-const { ipcRenderer } = require('electron');
+import { config } from './cameraConfig.js'
+import Cam from './cam.js'
 
-import cameraConfig from './cameraConfig.js';
-import Cam from './cam.js';
+navigator.mediaDevices.getUserMedia({
+  video: config || true
+}).then(stream => {
+  Cam.video.srcObject = stream
+})
 
-navigator.mediaDevices.enumerateDevices().then((devices) => {
-  const videoDevices = devices.filter((device) => {
-    return device.kind === 'videoinput';
-  });
+// Start Controlling Cam
+Cam.init()
 
-  // Using webcam in browser
-  navigator.mediaDevices
-    .getUserMedia({
-      video: { ...cameraConfig, deviceId: videoDevices[0].deviceId } || true
-    })
-    .then((stream) => (Cam.video.srcObject = stream));
+// Keyboard Events
+window.addEventListener('keydown', (e) => {
+  if (Cam.controls[e.key]) {
+    Cam.controls[e.key]()
+  }
+})
 
-  // Start Controlling Cam
-  Cam.init();
+/* Mac Only: Change Size */
+let isBig = false // the window starts small
 
-  ipcRenderer.send('videoInput', JSON.stringify(videoDevices));
-  ipcRenderer.on('videoInputChange', (event, videoInputIndex) => {
-    navigator.mediaDevices
-      .getUserMedia({
-        video:
-          {
-            ...cameraConfig,
-            deviceId: videoDevices[videoInputIndex].deviceId
-          } || true
-      })
-      .then((stream) => (Cam.video.srcObject = stream));
-  });
+window.ondblclick = changeWrapperSize
 
-  // Keyboard Events
-  window.addEventListener('keydown', (e) => {
-    if (Cam.controls[e.key]) {
-      Cam.controls[e.key]();
-    }
-  });
-});
+function changeWrapperSize () {
+  const { MiniVideoMe } = window
+
+  MiniVideoMe.sendDoubleClick(isBig)
+
+/* tested only on Mac */
+const hideCursor = {
+  timer: null,
+  mousemove: (e) => {
+    document.body.style.cursor = 'default'
+    clearTimeout(hideCursor.timer)
+    hideCursor.timer = setTimeout(() => {
+      document.body.style.cursor = 'none'
+    }, 1000)
+  }
+}
+
+window.addEventListener('mousemove', hideCursor.mousemove)
+/* end tested only on Mac */
