@@ -1,13 +1,16 @@
-export default {
-    width: 1920,
-    height: 1080,
-    aspectRatio: 1.77778,
-    frameRate: 59.94,
-    flipHorizontal: false, // start flipped
-    rounded: true, // border radius for camera
-    scale: 1.1, // zoom?
-    // move point of interest
-    // number will be converted to percent
-    horizontal: "-20", // 0 left - translateX
-    vertical: "0", // 0 top - translateY
+const userPreferences = window.MiniVideoMe.config
+
+console.log(userPreferences)
+
+const config = {
+  width: userPreferences.camera.width,
+  height: userPreferences.camera.height,
+  frameRate: userPreferences.camera.frameRate,
+  flipHorizontal: userPreferences.flipHorizontal,
+  rounded: userPreferences.rounded,
+  scale: userPreferences.zoom,
+  horizontal: String(userPreferences.anchor.x),
+  vertical: String(userPreferences.anchor.y)
 }
+
+export { config }
